+++ src/OTS.Data/OTS.Data.csproj
@@ -9,7 +9,6 @@
 
   <ItemGroup>
     <Folder Include="Interfaces\" />
-    <Folder Include="ViewModels\" />
     <Folder Include="Repositories\" />
   </ItemGroup>
