+++ src/OTS.Data/ViewModels/SubmitRelatedViewModel.cs
@@ -1,0 +1,16 @@
+﻿namespace OTS.Data.ViewModels
+{
+    public class SubmitViewModel
+    {
+        public Guid SubmitId { get; set; }
+        public Guid UserId { get; set; }
+        public Guid TestId { get; set; }
+        public double Score { get; set; }
+        public ICollection<SubmittedAnswerViewModel> SubmittedAnswerViewModels { get; set; } = new List<SubmittedAnswerViewModel>();
+    }
+    public class SubmittedAnswerViewModel
+    {
+        public Guid SubmittedAnswerId { get; set; }
+        public AnswerViewModel? Answers { get; set; }
+    }
+}
